Move DocumentLocaleSettings to its own file
