+++ lib/common.js
@@ -1,3 +1,5 @@
+import { DocumentLocaleSettings } from './document-locale-settings.js';
+
 export const defaultLocale = 'en';
 export const supportedBaseLocales = ['ar', 'cy', 'da', 'de', 'en', 'es', 'fr', 'haw', 'hi', 'ja', 'ko', 'mi', 'nl', 'pt', 'sv', 'tr', 'zh'];
 export const supportedLangpacks = ['ar', 'cy', 'da', 'de', 'en', 'en-gb', 'es', 'es-es', 'fr', 'fr-fr', 'fr-on', 'haw', 'hi', 'ja', 'ko', 'mi', 'nl', 'pt', 'sv', 'tr', 'zh-cn', 'zh-tw'];
@@ -87,140 +89,6 @@ export function validateFormatValue(value) {
 		throw new RangeError('value is out of range.');
 	}
 	return value;
-}
-
-class DocumentLocaleSettings {
-
-	constructor() {
-		this._cache = new Map();
-		this._htmlElem = document.documentElement;
-		this._listeners = [];
-		this._overrides = {};
-		this._observer = new MutationObserver(this._handleObserverChange.bind(this));
-		this._observer.observe(this._htmlElem, { attributes: true });
-		this.sync();
-		this._languageInitial = this._language;
-	}
-
-	get fallbackLanguage() { return this._fallbackLanguage; }
-	set fallbackLanguage(val) {
-		const normalized = this._normalize(val);
-		if (normalized === this._fallbackLanguage) return;
-		this._cache.clear();
-		this._fallbackLanguage = normalized;
-		this._listeners.forEach((cb) => cb());
-	}
-
-	get language() { return this._language; }
-	set language(val) {
-		const normalized = this._normalize(val);
-		if (normalized === this._language) return;
-		this._cache.clear();
-		this._language = normalized;
-		this._listeners.forEach((cb) => cb());
-	}
-
-	get overrides() { return this._overrides; }
-	set overrides(val) {
-		if (val.date) {
-			if (val.date.calendar) {
-				delete val.date.calendar.dayPeriods;
-			}
-			if (val.date.formats) {
-				delete val.date.formats.timeFormats;
-			}
-		}
-		this._cache.clear();
-		this._overrides = val;
-		this._listeners.forEach((cb) => cb());
-	}
-
-	addChangeListener(cb) {
-		this._listeners.push(cb);
-	}
-
-	getCacheItem(key, provider) {
-		if (!this._cache.has(key)) {
-			this._cache.set(key, provider());
-		}
-		return this._cache.get(key);
-	}
-
-	removeChangeListener(cb) {
-		const index = this._listeners.indexOf(cb);
-		if (index < 0) return;
-		this._listeners.splice(index, 1);
-	}
-
-	reset() {
-		this._cache.clear();
-		this._language = this._languageInitial;
-		this._fallbackLanguage = null;
-		this.overrides = {};
-		this.timezone = { name: '', identifier: '' };
-		this._listeners = [];
-		this.pseudoLocalization = { isAvailable: true, isEnabledForUser: false, textFormat: '{0}' };
-		this.oslo = { batch: null, collection: null, version: null };
-	}
-
-	sync() {
-		this.language = this._htmlElem.getAttribute('lang');
-		this.fallbackLanguage = this._htmlElem.getAttribute('data-lang-default');
-		this.overrides = this._tryParseHtmlElemAttr('data-intl-overrides', {});
-		this.timezone = this._tryParseHtmlElemAttr('data-timezone', { name: '', identifier: '' });
-		this.pseudoLocalization = this._tryParseHtmlElemAttr('data-pseudo-localization', { isAvailable: false, isEnabledForUser: false, replaceString: '', textFormat: '{0}' });
-		this.oslo = this._tryParseHtmlElemAttr('data-oslo', { batch: null, collection: null, version: null });
-	}
-
-	_handleObserverChange(mutations) {
-		let localeAttributeChange = false;
-		for (let i = 0; i < mutations.length; i++) {
-			const mutation = mutations[i];
-			if (mutation.attributeName === 'lang') {
-				this.language = this._htmlElem.getAttribute('lang');
-			} else if (mutation.attributeName === 'data-lang-default') {
-				this.fallbackLanguage = this._htmlElem.getAttribute('data-lang-default');
-			} else if (mutation.attributeName === 'data-intl-overrides') {
-				this.overrides = this._tryParseHtmlElemAttr('data-intl-overrides', {});
-			} else if (mutation.attributeName === 'data-timezone') {
-				this.timezone = this._tryParseHtmlElemAttr('data-timezone', { name: '', identifier: '' });
-				localeAttributeChange = true;
-			} else if (mutation.attributeName === 'data-pseudo-localization') {
-				this.pseudoLocalization = this._tryParseHtmlElemAttr('data-pseudo-localization', { isAvailable: false, isEnabledForUser: false, replaceString: '', textFormat: '{0}' });
-			} else if (mutation.attributeName === 'data-oslo') {
-				this.oslo = this._tryParseHtmlElemAttr('data-oslo', { batch: null, collection: null, version: null });
-				localeAttributeChange = true;
-			}
-		}
-		if (localeAttributeChange) this._listeners.forEach((cb) => cb());
-	}
-
-	_normalize(langTag) {
-
-		if (langTag === undefined || langTag === null) return null;
-
-		langTag = langTag.trim().toLowerCase();
-
-		const subtags = langTag.split('-');
-		if (subtags.length < 2) {
-			return langTag;
-		}
-
-		return `${subtags[0]}-${subtags[subtags.length - 1]}`;
-
-	}
-
-	_tryParseHtmlElemAttr(attrName, defaultValue) {
-		if (this._htmlElem.hasAttribute(attrName)) {
-			try {
-				return JSON.parse(this._htmlElem.getAttribute(attrName));
-			} catch {
-				// swallow exception
-			}
-		}
-		return defaultValue;
-	}
-
 }
 
 let documentLocaleSettings = null;
